httplog: move log writer to a separate file
This improves package structure.
